% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_column_metadata.R
\name{adm_column_metadata}
\alias{adm_column_metadata}
\title{Get column metadata}
\usage{
adm_column_metadata(database, server, table)
}
\arguments{
\item{database}{\code{string}. The database to get metadata data for.}

\item{server}{\code{string}. The server holding the database.}

\item{table}{\code{string}. The table to get metadata data for.}
}
\value{
\code{dataframe}
}
\description{
\code{adm_column_metadata} returns details of all columns in a SQL Server table.
}
\examples{

\dontrun{
adm_column_metadata(database = "DatabaseName", server = "ServerName", table = "TableName")
}

}
