% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_create_version.R
\name{adm_create_version}
\alias{adm_create_version}
\title{Create versioned dataset and record}
\usage{
adm_create_version(database, server, version_description, version_tables)
}
\arguments{
\item{database}{\code{string}. The database to get metadata data for.}

\item{server}{\code{string}. The server holding the database.}

\item{version_description}{\code{string}. The name of the version.}

\item{version_tables}{\code{vector}. A vector of staging tables used in the version.}
}
\value{
\code{NULL}
}
\description{
\code{adm_create_version} populates SQL Server temporal tables for all specified tables and records effective time.
}
\examples{

\dontrun{
adm_create_version(database = "DatabaseName", server = "ServerName", version_description = "ExampleVersionName", version_tables = c("Table1", "Table2"))
}

}
