% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_dataframe_to_db.R
\name{adm_dataframe_to_db}
\alias{adm_dataframe_to_db}
\title{Write dataframe to SQL Server database}
\usage{
adm_dataframe_to_db(
  database,
  server,
  database_table_name,
  dataframe,
  overwrite = FALSE,
  append = FALSE
)
}
\arguments{
\item{database}{\code{string}. The name of the database to write the sheet to.}

\item{server}{\code{string}. The name of the server containing the database.}

\item{database_table_name}{\code{string}. The name of the table to uplaod to in the database.}

\item{dataframe}{\code{string}. The dataframe to upload.}

\item{overwrite}{\code{string}. Default: \code{false}. Whether to overwrite existing table with worksheet with the same name.}

\item{append}{\code{string}. Default: \code{false}. Whether to append worksheet to existing table with the same name.}
}
\value{
\code{null}
}
\description{
\code{adm_dataframe_to_db} writes a dataframe to a SQL Server database.
}
\examples{

\dontrun{
adm_dataframe_to_db(database = "DatabaseName", server = "ServerName", database_table_name = "example_database_table", dataframe = example_dataframe)
}

}
