% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_delete_table.R
\name{adm_delete_table}
\alias{adm_delete_table}
\title{Delete a table in the 'dbo' schema from a database}
\usage{
adm_delete_table(database, server, table)
}
\arguments{
\item{database}{\code{string}. The database to delete the table from.}

\item{server}{\code{string}. The server holding the database.}

\item{table}{\code{string}. The table to be deleted.}
}
\value{
\code{null}
}
\description{
\code{adm_delete_table} opens an ODBC connection and deletes a specified table.
}
\examples{

\dontrun{
adm_delete_table(database = "DatabaseName", server = "ServerName", table = "TableName")
}

}
