% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_import_table.R
\name{adm_import_table}
\alias{adm_import_table}
\title{Read a SQL Server Table into R}
\usage{
adm_import_table(database, server, table, version = FALSE, date = NULL)
}
\arguments{
\item{database}{\code{string}. The database containing the table.}

\item{server}{\code{string}. The server holding the database.}

\item{table}{\code{string}. The table to read data from.}

\item{version}{\code{boolean}. Whether to retrieve the published table (as opposed to the staging table). Defaults to \code{FALSE}.}

\item{date}{\code{boolean}. The date to use to retrieve a versioned table. Defaults to \code{NULL}.}
}
\value{
\code{null}
}
\description{
\code{adm_import_table} reads the contents of a SQL Server database table to R.
}
\examples{

\dontrun{
adm_import_table(database = "DatabaseName", server = "ServerName", table = "TableName", version = FALSE)
}

}
