% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_import_version.R
\name{adm_import_version}
\alias{adm_import_version}
\title{Return all tables used in a version in a list}
\usage{
adm_import_version(database, server, version_description)
}
\arguments{
\item{database}{\code{string}. The database to get data from.}

\item{server}{\code{string}. The server holding the database.}

\item{version_description}{\code{string}. The description of the version as in [metadata].[Versions]}
}
\value{
\code{NULL}
}
\description{
\code{adm_import_version} returns all tables as were for a specific version.
}
\examples{

\dontrun{
adm_import_version(database = "DatabaseName", server = "ServerName", version_description = "Example version description")
}

}
