% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_list_versions.R
\name{adm_list_versions}
\alias{adm_list_versions}
\title{Return list of versions in version table}
\usage{
adm_list_versions(database, server)
}
\arguments{
\item{database}{\code{string}. A SQL Server database name.}

\item{server}{\code{string}. A SQL Server database server.}
}
\value{
\code{dataframe}
}
\description{
\code{adm_list_versions} lists all base tables in a SQL Server database on a specified server.
}
\examples{

\dontrun{
adm_list_versions(database = "DatabaseName", server = "DatabaseServer")
}

}
