% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_metadata_columns.R
\name{adm_metadata_columns}
\alias{adm_metadata_columns}
\title{Get column metadata}
\usage{
adm_metadata_columns(database, server, table)
}
\arguments{
\item{database}{\code{string}. The database to get metadata data for.}

\item{server}{\code{string}. The server holding the database.}

\item{table}{\code{string}. The table to get metadata data for.}
}
\value{
\code{dataframe}
}
\description{
\code{adm_metadata_columns} returns details of all columns in a SQL Server table.
}
\examples{

\dontrun{
adm_metadata_columns(database = "DatabaseName", server = "ServerName", table = "TableName")
}

}
