% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_metadata_tables.R
\name{adm_metadata_tables}
\alias{adm_metadata_tables}
\title{Get table metadata}
\usage{
adm_metadata_tables(database, server)
}
\arguments{
\item{database}{\code{string}. The database to get metadata data for.}

\item{server}{\code{string}. The server holding the database.}
}
\value{
\code{dataframe}
}
\description{
\code{adm_metadata_tables} returns details of all tables in a SQL Server database.
}
\examples{

\dontrun{
adm_metadata_tables(database = "DatabaseName", server = "ServerName")
}

}
