% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_rename_table.R
\name{adm_rename_table}
\alias{adm_rename_table}
\title{Rename SQL server database table}
\usage{
adm_rename_table(database, server, current_name, new_name)
}
\arguments{
\item{database}{\code{string}. The name of the database containing the table.}

\item{server}{\code{string}. The name of the server containing the database.}

\item{current_name}{\code{string}. The name of the existing table to rename.}

\item{new_name}{\code{string}. The name to rename the table to.}
}
\value{
\code{null}
}
\description{
\code{adm_rename_table} renames a table in a SQL Server database.
}
\examples{

\dontrun{
adm_rename_table(database = "DatabaseName", server = "ServerName", current_name = "table_name", new_name = "table_name_new")
}

}
