% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_table_metadata.R
\name{adm_table_metadata}
\alias{adm_table_metadata}
\title{Get table metadata}
\usage{
adm_table_metadata(database, server)
}
\arguments{
\item{database}{\code{string}. The database to get metadata data for.}

\item{server}{\code{string}. The server holding the database.}
}
\value{
\code{dataframe}
}
\description{
\code{adm_table_metadata} returns details of all tables in a SQL Server database.
}
\examples{

\dontrun{
adm_table_metadata(database = "DatabaseName", server = "ServerName")
}

}
