% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_upload_dataframe.R
\name{adm_upload_dataframe}
\alias{adm_upload_dataframe}
\title{Write dataframe to SQL Server database}
\usage{
adm_upload_dataframe(
  database,
  server,
  table,
  dataframe,
  overwrite = FALSE,
  append = FALSE
)
}
\arguments{
\item{database}{\code{string}. The name of the database to write the sheet to.}

\item{server}{\code{string}. The name of the server containing the database.}

\item{table}{\code{string}. The name of the table to upload to in the database.}

\item{dataframe}{\code{string}. The dataframe to upload.}

\item{overwrite}{\code{string}. Default: \code{false}. Whether to overwrite existing table with worksheet with the same name.}

\item{append}{\code{string}. Default: \code{false}. Whether to append worksheet to existing table with the same name.}
}
\value{
\code{null}
}
\description{
\code{adm_upload_dataframe} writes a dataframe to a SQL Server database.
}
\examples{

\dontrun{
adm_upload_dataframe(database = "DatabaseName", server = "ServerName", table = "example_database_table", dataframe = example_dataframe)
}

}
