% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_upload_folder.R
\name{adm_upload_folder}
\alias{adm_upload_folder}
\title{Upload all Excel worksheets in a folder to SDR}
\usage{
adm_upload_folder(
  database,
  server,
  source_directory_path,
  append = FALSE,
  overwrite = FALSE,
  archive_directory_path = NULL
)
}
\arguments{
\item{database}{\code{string}. The database to write the data to.}

\item{server}{\code{string}. The server holding the database.}

\item{source_directory_path}{\code{string}. The path of the directory containing Excel files.}

\item{append}{\code{string}. Default: \code{false}. Whether to append worksheets to existing tables with the same name.}

\item{overwrite}{\code{string}. Default: \code{false}. Whether to overwrite existing tables with worksheets with the same name.}

\item{archive_directory_path}{\code{string}. The path of the directory to send processed files to.}
}
\value{
\code{null}
}
\description{
\code{adm_upload_folder} uploads all worksheets in all Excel workbooks in a specified location to a specified database. Processed folders are then moved from the source_directory_path folder to an archive folder.
}
\examples{

\dontrun{
adm_upload_folder(database = "DatabaseName", server = "ServerName\\\\Instance", file_path = "C:\\\\Temp\\\\ExcelFile.xlsx")
}
\dontrun{
adm_upload_folder(database = "DatabaseName", server = "ServerName\\\\Instance", file_path = "C:\\\\Temp\\\\ExcelFile.xlsx", archive_directory_path = "C:\\\\ArchiveFolder\\\\")
}

}
