% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_upload_workbook.R
\name{adm_upload_workbook}
\alias{adm_upload_workbook}
\title{Process files to write to SQL Server database}
\usage{
adm_upload_workbook(
  database,
  server,
  file_path,
  append = FALSE,
  overwrite = FALSE,
  archive = NULL
)
}
\arguments{
\item{database}{\code{string}. The name of the database to write the file to.}

\item{server}{\code{string}. The name of the server containing the database.}

\item{file_path}{\code{string}. The full path of the file to process.}

\item{append}{\code{string}. Default: \code{false}. Whether to append worksheets to existing tables with the same name.}

\item{overwrite}{\code{string}. Default: \code{false}. Whether to overwrite existing tables with worksheets with the same name.}

\item{archive}{\code{string}. Optional. The path of the directory to move the processed file to.}
}
\value{
\code{null}
}
\description{
\code{adm_upload_workbook} writes all worksheets in an Excel workbook to a SQL Server database.
}
\examples{

\dontrun{
adm_upload_workbook(database = "DatabaseName", server = "ServerName\\\\Instance", file_path = "C:\\\\Temp\\\\ExcelFile.xlsx")
}
\dontrun{
adm_upload_workbook(database = "DatabaseName", server = "ServerName\\\\Instance"file_path = "C:\\\\Temp\\\\ExcelFile.xlsx", archive = "C:\\\\ArchiveFolder\\\\")
}

}
