% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_workbook_to_db.R
\name{adm_workbook_to_db}
\alias{adm_workbook_to_db}
\title{Process files to write to SQL Server database}
\usage{
adm_workbook_to_db(database, server, file_path, archive = NULL)
}
\arguments{
\item{database}{\code{string}. The name of the database to write the file to.}

\item{server}{\code{string}. The name of the server containing the database.}

\item{file_path}{\code{string}. The full path of the file to process.}

\item{archive}{\code{string}. Optional. The path of the directory to move the processed file to.}
}
\value{
\code{null}
}
\description{
\code{adm_workbook_to_db} writes all worksheets in an Excel workbook to a SQL Server database.
}
\examples{

\dontrun{
adm_workbook_to_db(database = "DatabaseName", server = "ServerName\\\\Instance", file_path = "C:\\\\Temp\\\\ExcelFile.xlsx")
}
\dontrun{
adm_workbook_to_db(database = "DatabaseName", server = "ServerName\\\\Instance"file_path = "C:\\\\Temp\\\\ExcelFile.xlsx", archive = "C:\\\\ArchiveFolder\\\\")
}

}
