% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_worksheet_to_db.R
\name{adm_worksheet_to_db}
\alias{adm_worksheet_to_db}
\title{Process worksheet from Excel file to write to SQL Server database}
\usage{
adm_worksheet_to_db(database, server, file_path, worksheet)
}
\arguments{
\item{database}{\code{string}. The name of the database to write the sheet to.}

\item{server}{\code{string}. The name of the server containing the database.}

\item{file_path}{\code{string}. The full path of the file containing the worksheet.}

\item{worksheet}{\code{string}. The name of the worksheet to be processed.}
}
\value{
\code{null}
}
\description{
\code{adm_worksheet_to_db} is an internal function that processes an Excel worksheet to be written to a SQL Server database.
}
\examples{

\dontrun{
adm_worksheet_to_db(database = "DatabaseName", server = "ServerName", file_path = "C:\\\\Temp\\\\ExcelFile.xlsx", worksheet = "WorksheetName")
}

}
